restructure attestation feature
