--- frontend/src/features/attestation/AttestationHeader.tsx
@@ -1,7 +1,0 @@
-export function AttestationHeader() {
-    return (
-        <header className="flex h-1/12 items-center bg-teal-900/40 p-3">
-            {/* <p className="text-2xl text-teal-50">Attestation</p> */}
-        </header>
-    );
-}
