+++ frontend/src/features/attestation/AttestationPage.tsx
@@ -3,10 +3,9 @@ import {
     // AttestationHeader,
     AttestationSummary,
     AttestationTimeline,
-    ImageVerification,
-} from '../features/attestation';
+} from '.';
 
-export function Attestation() {
+export function AttestationPage() {
     return (
         <main className="flex flex-1 flex-col space-y-3 bg-gradient-to-r from-teal-950 from-2% to-slate-800 to-95%">
             {/* <AttestationHeader /> */}
@@ -14,7 +13,6 @@ export function Attestation() {
             {/* TODO: Move Evidence in modal window? */}
             {/* <AttestationEvidence /> */}
             <AttestationTimeline />
-            <ImageVerification />
         </main>
     );
 }
