--- frontend/src/features/attestation/ImageVerification.tsx
@@ -1,3 +1,0 @@
-export function ImageVerification() {
-    return <div></div>;
-}
