+++ frontend/src/features/attestation/attestationAPI.ts
@@ -3,6 +3,12 @@ const ATTESTER_URL = import.meta.env.VITE_ATTESTER_URL;
 // TODO: Refactor using Axios.
 // TODO: Add error handling.
 
+/**
+ * Function that communicates with a third-party evidence provider service. It retrieves an attestation quote from a TDX TEE.
+ *
+ * @param challenge
+ * @returns a valid TDX attestation quote.
+ */
 export async function getTdxQuote(challenge: string) {
     try {
         const response = await fetch(`${ATTESTER_URL}/evidence/tdx-quote`, {
