+++ frontend/src/features/attestation/components/Evidence/TeeEvidence.tsx
@@ -1,10 +1,10 @@
 import { useMutation } from '@tanstack/react-query';
 import { JsonEditor } from 'json-edit-react';
 
-import { attestationEvidence } from '../../dev-data/attestation-evidence';
-import { getTdxQuote } from '../../services/attester';
+import { attestationEvidence } from '../../../../dev-data/attestation-evidence';
+import { getTdxQuote } from '../../attestationAPI';
 
-export function AttestationEvidence() {
+export function TeeEvidence() {
     const mutation = useMutation({ mutationFn: getTdxQuote });
 
     return (
