+++ frontend/src/features/attestation/components/Evidence/WorkloadEvidence.tsx
@@ -1,0 +1,1 @@
+export function WorkloadEvidence() {}
