+++ frontend/src/features/attestation/components/Evidence/index.ts
@@ -1,0 +1,2 @@
+export * from './TeeEvidence';
+export * from './WorkloadEvidence';
