+++ frontend/src/features/attestation/components/Forms/index.ts
@@ -1,0 +1,1 @@
+export * from './IssueChallengeForm';
