+++ frontend/src/features/attestation/components/Summary/index.ts
@@ -1,0 +1,1 @@
+export * from './AttestationSummary';
