+++ frontend/src/features/attestation/components/Timeline/index.ts
@@ -1,0 +1,2 @@
+export * from './AttestationStep';
+export * from './AttestationTimeline';
