+++ frontend/src/features/attestation/components/WorkloadIntegrity/WorkloadIntegrity.tsx
@@ -1,0 +1,1 @@
+export function WorkloadIntegrity() {}
