+++ frontend/src/features/attestation/components/WorkloadIntegrity/index.ts
@@ -1,0 +1,1 @@
+export * from './WorkloadIntegrity';
