+++ frontend/src/features/attestation/index.ts
@@ -1,6 +1,9 @@
-export * from './AttestationHeader';
-export * from './AttestationSummary';
-export * from './AttestationTimeline';
-export * from './AttestationEvidence';
-export * from './ImageVerification';
-export * from './IssueChallengeForm';
+export * from './components/Summary';
+export * from './components/Timeline/';
+export * from './components/WorkloadIntegrity';
+export * from './components/Forms/IssueChallengeForm';
+export * from './components/Evidence';
+export * from './AttestationPage';
+export * from './attestationAPI';
+export * from './attestationSlice';
+export { default as attestationSlice } from './attestationSlice';
