+++ frontend/src/pages/index.ts
@@ -2,5 +2,4 @@ export * from './AppLayout';
 export * from './Auth';
 export * from './Chat';
 export * from './User';
-export * from './Attestation';
 export * from './Overview';
